feat: convert TextCopyField to TypeScript
